src/main/java/com/resume/resume/repository/UserRepository.java
package com.resume.resume.repository;

import com.resume.resume.model.Role;
import com.resume.resume.model.User;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;
import org.springframework.stereotype.Repository;

import javax.swing.text.html.Option;
import java.util.Optional;

@Repository
public interface UserRepository extends JpaRepository<User, Long> {

    @Query("SELECT u FROM User u WHERE u.email = ?1")
    public Optional<User> findByEmail(String email);

    @Query("SELECT u FROM User u WHERE u.role = ?1")
    public Optional<User> findByRole(Role role);
}
